map fog, Object Oriented Village, and fight
I implemented the map fog feature, so the player only see the surrounding area.

The Village has been updated to OOP. This is the first step forward to the transitioning from functional programming to OOP. 
This update causes minor bugs that can't be avoided. But because the whole program will be written in OOP these will disappear.

Now there is simple fight mechanism in the game that utilizes 2 new features: Health point and Mana point.
There is a 3% chance that in Bushes a Prowling Tiger or in Jungles a Jungle Troll attacks the player.